tutorial/pyproject.toml
[tool.poetry]
name = "tutorial"
version = "0.1.0"
description = ""
authors = ["Marco Enrico <mvalviar@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.6"
djangorestframework = "^3.15.1"
django-cors-headers = "^4.3.1"
django-todo = "^2.5.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
